Repository set up for Sphinx HTML generation
The format of this initial setup does not include any files generated by
Sphinx or even any files describing myself. It just gets the repository
organized for running Sphinx on the documents that will be created.

This setup was based on the original CRACMM documentation website
releaced in March of 2025 and was still the setup at the time of
writing. For reference, enter the following link into your browser:
https://github.com/USEPA/CRACMM/tree/CRACMMdocs

Changes to be committed:
	new file:   ../conf.py
	new file:   ../index.md
	new file:   Makefile
	new file:   _static/headshot.jpg
	new file:   make.bat
	new file:   python_scripts/add_toctree.py
	new file:   python_scripts/delete_toctree.py
	new file:   run_build.csh
	new file:   toctree.txt